apps/api/pyproject.toml
[project]
name = "student-planner-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "firebase-admin>=7.1.0",
    "httpx>=0.28.1",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langsmith>=0.6.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
